+++ src/bump_string.rs
@@ -580,7 +580,7 @@ where
     /// Shrinks the capacity of the string as much as possible.
     ///
     /// This will also free space for future bump allocations iff this is the most recent allocation.
-    /// 
+    ///
     /// # Examples
     /// ```
     /// # use bump_scope::{ Bump, BumpString };