+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **added**: `from_utf8_unchecked` for all string types
+
 ## 0.5.6 (2024-08-17)
 - **fixed:** `alloc_iter` and `alloc_fmt` to not take up more than the necessary memory